Delete all *.pyc files and \_\_pycache\_\_ folders
Python automatically generates __pycache__ and *.pyc files whenever a
Python file is executed. If these files and folders are committed to the
repository, they are shown as changes whenever they are pulled to the
local repository. Thus, they should be removed.

Note: since *.pyc files are cache files (not source files), removing
them do not impact the functionality of the project.